Move a test in a `Task` method to its own test module
